refactor(ts-render-tauri): move VideoCompilerState + ServiceContainer + services into crate (Wave 1, #91)
Перенос shared-state-фундамента video_compiler в крейт ts-render-tauri
(Option B, волна 1). Командные группы (26 шт.) НЕ трогаются — пути сохранены
через ре-экспорт-шимы.

Перенесено (git mv, логика не менялась):
- video_compiler/services/* (8 сервисов + ServiceContainer + ServiceMetrics)
  -> crates/ts-render-tauri/src/services/
- video_compiler/commands/state.rs (VideoCompilerState/ActiveRenderJob/
  RenderJobMetadata/RenderJob) -> crates/ts-render-tauri/src/state.rs
- core/ffmpeg/{mod.rs FFmpegCommand-примитив, security.rs}
  -> crates/ts-render-tauri/src/ffmpeg_command/ (preview_service зависит от
  FFmpegCommand::ffprobe; анализное поддерево остаётся в монолите)

Пути внутри перенесённых файлов переписаны: crate::video_compiler::services::*
-> crate::services::*; crate::video_compiler::{cache,progress,renderer,schema,
core::*,Result,CompilerSettings} -> ts_render::video_compiler::core::* (единый
источник истины — движок ts-render, схема ts-schema).

Ре-экспорт-шимы в монолите (нулевые правки у потребителей):
- video_compiler/services/mod.rs -> pub use ts_render_tauri::services::*
- video_compiler/commands/state.rs -> pub use ts_render_tauri::state::*
- core/ffmpeg/mod.rs -> pub use ts_render_tauri::ffmpeg_command::* (FFmpegCommand,
  FFmpegSecurity, check_*, GpuCapabilities для super::FFmpegCommand в 11 анализных
  подмодулях и внешнего unified_audio_analyzer)
- commands/mod.rs: #[allow(ambiguous_glob_reexports)] на cache::* (CacheStats
  теперь крейтовый тип, флэтовый commands::CacheStats нигде не используется)

B2 (критично): тип в tauri::Builder.manage() сохранён ТОЧНО. lib.rs не правился —
video_compiler::initialize() и VideoCompilerState::new() прозрачно строят крейтовый
тип через шимы. Единственное определение VideoCompilerState в сборке — крейтовое,
поэтому все 347 State<VideoCompilerState>-экстракторов резолвятся в один тип
(runtime State<T>-mismatch структурно невозможен).

Зависимости крейта: + ts-render, ts-schema, tokio, async-trait, uuid, once_cell,
chrono, log; dev: + futures, filetime.

Проверки: cargo check -p ts-render-tauri = ok; cargo test -p ts-render-tauri =
226 passed/0 failed; cargo check --lib (src-tauri) = ok (baseline 4 warnings,
без новых); cargo test --lib --no-run (src-tauri) = ok (моки/тесты компилируются
через шимы, включая struct-литерал VideoCompilerState).
