+++ crates/ts-render-tauri/src/services/ffmpeg_service.rs
@@ -1,9 +1,7 @@
 //! FFmpeg сервис для работы с видео
 
-use crate::video_compiler::{
-  error::{Result, VideoCompilerError},
-  services::Service,
-};
+use crate::services::Service;
+use ts_render::video_compiler::core::error::{Result, VideoCompilerError};
 use async_trait::async_trait;
 use std::{
   path::Path,
